Update to improve code readability
